IndexedDB/submissions/Microsoft/idbfactory_cmp.htm

<!DOCTYPE html>
<html>
<head>  
    <title>IDBFactory.cmp() - compared keys return correct value</title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript"> 
        function RunTest() {
            var greater = window.indexedDB.cmp(2, 1);
            var equal = window.indexedDB.cmp(2, 2);
            var less = window.indexedDB.cmp(1, 2);
            
            assert_equals(greater, 1, "greater");
            assert_equals(equal, 0, "equal");
            assert_equals(less, -1, "less");
        }
        
        test(RunTest, "IDBFactory.cmp()", null);
    </script>
</head>
<body>
    <div id="log">
    </div>
</body>
</html>
